Create Match and refactor Module
A kinda big change here is that the string id for modules are gone.
I'll probably change the concept of module entries tomorrow as well.
They're kinda bulky right now requiring a confusing and unnessisary
method. The annotation could just as easily be tacked on a class and
arbitrarily require a no-params constructor.