Add KitModule
Because this commit took me forever an a day to put together (despite being 50%
copy-paste), it has a handful of other additions. Notably, a bunch of added util
methods, and slightly revised module loading with the hope that it is more
predictable.

Due to Elly's modules being broken, I couldn't test this as much as I wanted,
but at least it doesn't throw any exceptions when loading.